+++ client/src/StoryList.tsx
@@ -1,12 +1,11 @@
-import _ from "lodash";
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { axiosInst } from ".";
-import { getCurrentFeeds } from "./AddFeed";
 import { timeSince } from "./helpers";
+import { getReadStoryList } from "./localStorage";
+import _ from "lodash";
 
-interface StoryListState {
+interface StoryListProps {
   stories: Story[];
 }
 
@@ -18,72 +17,48 @@ export interface Story {
   date: any;
 }
 
-const LOCAL_STORIES = "STORIES";
-export class StoryList extends React.Component<{}, StoryListState> {
-  constructor(props: {}) {
+export class StoryList extends React.Component<StoryListProps, {}> {
+  constructor(props: StoryListProps) {
     super(props);
-
-    this.state = { stories: [] };
   }
-  componentDidMount() {
-    const _stories = localStorage.getItem(LOCAL_STORIES);
-    if (_stories === null) {
-      return;
-    }
-
-    // these will be sorted
-    const stories = JSON.parse(_stories) as Story[];
-
-    this.setState({ stories });
-  }
+  componentDidMount() {}
   render() {
-    if (this.state.stories.length === 0) {
+    if (this.props.stories.length === 0) {
       return <div>refresh stories or add a feed to get started</div>;
     }
+
+    const readStories = getReadStoryList();
+
+    console.log("read stories", readStories);
 
     return (
       <div>
         <div>
-          {this.state.stories.map((story) => (
-            <div
-              style={{ display: "flex", alignItems: "center" }}
-              key={story.url}
-            >
-              <img
-                src={`https://s2.googleusercontent.com/s2/favicons?domain_url=${story.url}`}
-                style={{ flexGrow: 0 }}
-              />
-              <p key={story.url}>
-                <Link to={`/story/${encodeURIComponent(story.url)}`}>
-                  <b>{story.title}</b>
-                  <span> {timeSince(story.date)}</span>
-                </Link>
-              </p>
-            </div>
-          ))}
+          {this.props.stories.map((story) => {
+            const isReadStory = readStories[story.url];
+            return (
+              <div
+                style={{ display: "flex", alignItems: "center" }}
+                key={story.url}
+              >
+                <img
+                  src={`https://s2.googleusercontent.com/s2/favicons?domain_url=${story.url}`}
+                  style={{ flexGrow: 0 }}
+                />
+                <p key={story.url}>
+                  <Link
+                    to={`/story/${encodeURIComponent(story.url)}`}
+                    style={{ color: isReadStory ? "#686868" : undefined }}
+                  >
+                    <b>{story.title}</b>
+                    <span> {timeSince(story.date)}</span>
+                  </Link>
+                </p>
+              </div>
+            );
+          })}
         </div>
       </div>
     );
-  }
-  async refreshFeedContents() {
-    const feeds = getCurrentFeeds();
-
-    const feedRes = await axiosInst.post("/api/feed_update", { feeds });
-
-    console.log(feedRes);
-
-    // it's back, update stories
-
-    const stories = feedRes.data;
-
-    this.saveAndUpdateState(stories);
-  }
-
-  private saveAndUpdateState(stories: Story[]) {
-    const storiesSorted = _.sortBy(stories, (c) => -c.date);
-
-    localStorage.setItem(LOCAL_STORIES, JSON.stringify(storiesSorted));
-
-    this.setState({ stories: storiesSorted });
   }
 }